Move preview orders tests to orders tests
* dart format
