+++ example/get_btc_price_example.dart
@@ -15,7 +15,8 @@ void main() async {
   // Print the price to the console.
   if (ticker != null) {
     if (ticker.trades != null && ticker.trades!.isNotEmpty) {
-      print('The latest trade price of BTC-USD is: ${ticker.trades!.first.price}');
+      print(
+          'The latest trade price of BTC-USD is: ${ticker.trades!.first.price}');
     } else {
       print('No recent trades found for BTC-USD.');
     }
